requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
google-generativeai
python-dotenv
